+++ src/multilevelmenu.js
@@ -11,24 +11,14 @@
 
     'use strict';
 
-    var support = { animations : Modernizr.cssanimations },
-        animEndEventNames = { 'WebkitAnimation' : 'webkitAnimationEnd', 'OAnimation' : 'oAnimationEnd', 'msAnimation' : 'MSAnimationEnd', 'animation' : 'animationend' },
-        animEndEventName = animEndEventNames[ Modernizr.prefixed( 'animation' ) ],
-        onEndAnimation = function( el, callback ) {
-            var onEndCallbackFn = function( ev ) {
-                if( support.animations ) {
-                    if( ev.target !== this ) return;
-                    this.removeEventListener( animEndEventName, onEndCallbackFn );
-                }
-                if( callback && typeof callback === 'function' ) { callback.call(); }
-            };
-            if( support.animations ) {
-                el.addEventListener( animEndEventName, onEndCallbackFn );
-            }
-            else {
-                onEndCallbackFn();
-            }
+    var onEndAnimation = function( el, callback ) {
+        var onEndCallbackFn = function( ev ) {
+            if( ev.target !== this ) return;
+            this.removeEventListener( 'animationend', onEndCallbackFn );
+            if( callback && typeof callback === 'function' ) { callback.call(); }
         };
+        el.addEventListener( 'animationend', onEndCallbackFn );
+    };
 
     function extend( a, b ) {
         for( var key in b ) {