+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Restructured project 
+- Removed modernizr fallback for cssanimations
 
 ## [1.0.1] - 2017-11-22